+++ routers/router.go
@@ -4,12 +4,15 @@
 package routers
 
 import (
-	"github.com/udistrital/sga_syllabus_mid/controllers"
-
 	"github.com/astaxie/beego"
+	"github.com/udistrital/sga_syllabus_mid/controllers"
+	"github.com/udistrital/utils_oas/errorhandler"
 )
 
 func init() {
+
+	beego.ErrorController(&errorhandler.ErrorHandlerController{})
+
 	ns := beego.NewNamespace("/v1",
 		beego.NSNamespace("/syllabus",
 			beego.NSInclude(
